refactoring game.js
